chore(storage): migrate OC tracing test
Migrate this test to gRPC and move it into the main integration
test file.
